requirements.txt
openpyxl>=3.1.0
Pillow>=10.0.0
anthropic>=0.30.0
requests>=2.31.0
jinja2>=3.1.0
flask>=3.0.0
google-genai>=1.0.0
